move route utility functions into services.
